rs_correct_msg: report correct errata positions when using erasures
rs_correct_msg zeroes erased bytes before decoding, then takes an early
return if the message is correct, returning an empty list of errata
positions.  This leads to two inconsistencies:

- If the zeroing corrects the message, we would incorrectly return an
empty list of errata positions.

- If the input message was correct, the zeroing usually corrupts the
message, so we don't take the early return, and then we correct the
message and return the errata positions (including all erased
positions).  But if the erased bytes in the (correct) input message
happen to be zero, we take the early return and return an empty list of
errata positions.

This commit retains the zeroing and the early return, but fixes the
above inconsistencies by returning the list of erased positions in the
early return.